source/OAuth2OOo/requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.2
calmjs==3.4.4
calmjs.parse==1.3.1
calmjs.types==1.0.1
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset_normalizer==3.3.2
cssselect==1.2.0
decorator==5.1.1
exceptiongroup==1.2.0
extruct==0.16.0
h11==0.14.0
html_text==0.5.2
html5lib>=1.1
idna>=3.7
ijson>=3.2.3
isodate>=0.6.1
jmespath>=1.0.1
js2xml>=0.5.0
jsonpath_ng>=1.6.0
jstyleson>=0.0.2
lxml>=5.1.0
mf2py>=2.0.1
outcome>=1.3.0.post0
packaging>=23.2
parsel>=1.8.1
ply>=3.11
pycparser>=2.21
pyparsing>=3.1.1
pyRdfa3>=3.6.2
PySocks>=1.7.1
python-dotenv>=1.0.0
rdflib>=7.0.0
requests>=2.32.3
selenium>=4.16.0
setuptools>=69.0.3
six>=1.16.0
sniffio>=1.3.0
sortedcontainers>=2.4.0
soupsieve>=2.5
tqdm>=4.66.4
trio>=0.23.2
trio_websocket>=0.11.1
urllib3>=2.1.0
validators>=0.22.0
w3lib>=2.1.2
webdriver_manager>=4.0.1
webencodings>=0.5.1
wsproto>=1.2.0
